added Company class, edited how IDs work
Edited UserID class so it gave different IDs to different classes (PublicUser is P0000125 and Company is C0000125, and the counts are separate, as in the 4th company is C0000004 and the 4th PublicUser is P0000004). Has error message if the user type isn't PublicUser or Company. Added test to Main and it works.