More Order changes and tests
Added several Order changes, now an ArrayList of Orders is contained within every PublicUser and an order can be placed, returned (ie removed from the ArrayList), and the PublicUser's Orders can be returned as a list within a String or as an ArrayList. Placed Orders remain in the static Order.ORDERS_AND_THEIR_IDS HashMap after their completion, so in effect a backlog of every Order ever placed is kept.