Added Firebase
Authentication system is working, plus the beginnings of database - currently it contains users with same uid as the user system, with the additional value Postcode. May stretch to Address too in future.
LOTS OF CHANGES. Also added a "Test Town" section which for the moment links to a fake user with 4 orders already made or a Company user.
Also added the beginnings of a Company User Home Page, and them making choices on Orders as they come in.